requirements.txt
mysql-connector-python==8.0.18
selenium==3.141.0